python/gcp_cai/requirement.txt
google-cloud-asset
google-api-core